Correctly implement AlwaysReplace policy when condensing.
When condensing the slots in a set, if the condensation policy is
AlwaysReplace, this means that the values from the individual mappings
must always take precedence over any value that may exist at the set
level.

So we must deal with two cases:

A) A slot is condensable (all mappings in the set have the same value
for the slot) but the set has its own value (different from the values
of the mappings) for that slot. In that case, the set-level value is
replaced by the condensed value from the individual mappings.

B) A slot is not condensable (mappings have different values for that
slot), *and* the set also has its own value. In that case, the set-level
value is deleted.

This does not affect the TSV writer, since it uses the NeverReplace
policy -- which leaves the set-level values untouched. I am unsure as to
whether (1) it should use AlwaysReplace instead, and/or (2) that should
be configurable. For now, I will leave it like that.