Update OWLAPI dependency to 4.5.29.
Bump the OWLAPI dependency to the latest 4.5.29 release.

For the ROBOT plugin, this requires temporarily adding an explicit
dependency to a recent (>= 1.3) version of Logback-classic, because
ROBOT currently depends on a version (1.2.13) that is not compatible
with the SLF4J-API 2.0.11 that is now pulled by the new OWLAPI.

Without an explicit dependency to Logback-classic, when we build the
robot package for testing, we would end up with SLF4J-API 2.0.11 (as
requested by the OWLAPI) and Logback-classic 1.2.13 (as requested by
ROBOT), and the tests would fail because of the incompatibilities
between those two versions.

(Presumably, and somewhat ironically, the actual ROBOT plugin should
still work fine, since it does not bring its own dependencies but relies
instead on those provided by ROBOT, and the current ROBOT would bring
SLF4J-API 1.7.32 and Logback-classic 1.2.13, which are compatible.)

When ROBOT 1.10 will be released, we'll be able to rely again on the
inherited dependency to Logback-classic.