Only escape YAML strings if necessary.
Do not systematically write strings in YAML double-quoted style. Only do
so when the string is not suitable for the plain style (e.g. it starts
with a YAML indicator or contains characters that need to be escaped).

This is to align with SSSOM-Py behaviour, and with the proposed
"canonical" variant of the SSSOM/TSV format in the updated
specification.