Fix YAML escaping and add more tests.
Make sure to escape a string that starts with a white space -- otherwise
the space would be trimmed away when the file is read back.

Also fix the logic that checks whether an initial ':', '?', or '-'
character is followed by a white space.

Add various cases of strings that need (or not) escaping in the existing
YAML escaping test.