Make sure scalar values are correctly parsed.
The YAML parser will automatically try to parse some scalar values into
what it thinks is the most appropriate type of Java object. For example,
with an input like this:

  comment: 123

the parsed value will be an Integer object, not a string.

This may be nice for most people, but in our case we do not need that,
as we already have the LinkML schema and we always know exactly what
type to expect, without having to guess. In the above example, the
"comment" slot should be a string, not an integer. Arguably it may be
best for creators of SSSOM sets to always avoid using unquoted string
values if they may be mistaken for integer values, but we should allow
that all the same -- we should not require that such values have to be
quoted, precisely because we have the schema to tell us that the value
should be a string anyway, even if it _looks like_ an integer.

Unfortunately it is not possible (as far as I know) to configure the
Jackson YAML parser to prevent it from parsing "integer-looking" values
as integers, so we need to be ready, when analysing the output of the
parser, to accept variously typed objects, instead of only strings.