Finalise the YAML-to-MS conversion when reading metadata only.
When we read only the metadata section of a mapping set, make sure to
finalise the conversion of the parsed YAML objects into a proper
MappingSet object, by calling the `postMappings()` method of the
YAMLConverter. Otherwise, undefined extensions that may have been used
in the metadata block will not be taken into account.