Support writing external metadata.
This commit adds full support for writing SSSOM/TSV files in "external
metadata" mode, where the metadata block is written in a separate file
than the TSV section.

The TSVWriter class gets new constructors that accept a second argument
representing the file/stream/filename where the metadata block should be
written.

The SSSOM-CLI tool gets a new --metadata-output option to specify the
name of the file where to write the metadata block. Without that option,
the default behaviour is still to write the set in "embedded" mode.