Add the "Propagatable" annotation.
The LinkML model for SSSOM will soon include a "propagated" annotation
to explicitly mark the propagatable slots.

We transpose that LinkML annotation into a Java field annotation so that
propagatable fields in the MappingSet class are marked as such.

Because the default Java generator provided with LinkML does not seem to
give access to the list of annotations carried by a slot (or any other
schema element for that matter), we use a custom generator script that
provides a helper method to say whether a given slot has a "propagated"
annotation.