Change options that control the "update from ontology".
Change the OWLHelper.UpdateMode enum to remove the following values:

* DELETE_MISSING_SUBJECT,
* DELETE_OBSOLETE_SUBJECT,
* DELETE_MISSING_OBJECT,
* DELETE_OBSOLETE_OBJECT

and replace them with the following values instead:

* DELETE_MISSING,
* DELETE_OBSOLETE,
* SUBJECT_ONLY,
* OBJECT_ONLY.

The new logic is that the DELETE_MISSING and DELETE_OBSOLETE mode will
operate indistinctly on both the subject side and the object side of
mappings, unless either SUBJECT_ONLY or OBJECT_ONLY are also specified
to instruct the updater to only consider one side.

SUBJECT_ONLY and OBJECT_ONLY also apply to the other update modes
(UPDATE_LABEL and UPDATE_SOURCE). For example, with a mode set to
SUBJECT_ONLY|UPDATE_LABEL, only the subject label will ever be updated,
even if the object also exists in the ontology.

The change is reflected in SSSOM-CLI's --update-from-ontology option,
which now accepts 3 flags to dictate what should be updated:

* label (update the labels),
* source (update the sources),
* existence (delete mapping if the subject or object does not exist);

and 2 flags to dictate whether to check/update only the object or only
the subject:

* subject (only update subject's label and source, and delete if subject
  is missing);
* object (only update object's label and source, and delete if object is
  missing).