Add basic JSONWriter.
We start adding support for JSON format by adding a JSONWriter.

The generated JSON is merely a direct JSON rendering of the internal
SSSOM data model. In particular, it is NOT JSON-LD! As such, the
produced files cannot be read by SSSOM-Py, because what SSSOM-Py calls
"json" is really JSON-LD -- it does not expect "pure JSON".