Add basic JSON reader.
We add a new JSONReader class to implement the "pure" JSON format in
read mode. The reader is particuarly simple as most of the logic to deal
with converting the raw JSON objects into SSSOM objects is already
implemented in the YAMLConverter class (which is actually slightly
misnamed, as it is independent of YAML; oh well).

As for the writer objects, we use a new BaseReader class to factor code
that is common to both the TSVReader and the JSONReader, even though
this is not as important as for the writers as there is much less common
code (since, again, most of the logic that is relevant to both readers
is in the YAMLConverter).