Support SSSOM-Py JSON-LD output.
We add a new YAMLPreprocessor that can turn the JSON-LD produced by
SSSOM-Py into a form that can be accepted by the YAMLConverter. This
preprocessor does two things:

* it strips away any `@type` key (we don't need them, and such a key
  could be interpreted as a SSSOM "undefined extension slot");
* it extracts prefix declarations from the "@context" key and puts them
  into a proper SSSOM `curie_map` slot.

This allows the JSONReader to accept the JSON files that are produced by
SSSOM-Py.

Conversely, we add a new option to the JSONWriter class to force it to
write the curie map in a JSON-LD-like "@context" dictionary. This is
enough to make the resulting JSON accepted by SSSOM-Py.