Support reading both SSSOM/TSV and SSSOM/JSON.
Since the SSSOM/TSV reader already has to peek into the TSV file to
detect whether it has an embedded metadata block or not (i.e., check if
the first byte is a '#'), we can reuse that logic to also detect if the
file is a SSSOM/JSON file instead (if the first byte is a '{').

This allows to read indifferently SSSOM/TSV and SSSOM/JSON files using
the same code: if you don't know whether the file you have to read is a
SSSOM/TSV file or a SSSOM/JSON file, just use a TSVReader and it will
automatically do The Right Thing™. Notably, this automatically gives
SSSOM-CLI the ability to read JSON files.

(If you do know that you will be reading a JSON file, you can -- and
should -- still explicitly use a JSONReader, of course.)