Add JSON output to SSSOM-CLI.
Add the possibility to write the result mapping set in JSON format
instead of SSSOM/TSV format. This is controlled by several boolean
options:

  -j, --json-output: Enable JSON output.

  --json-short-iris: Shorten IRIs in JSON output. The default is to
  write IRIs in their full-length form. That option is only meaningful
  when --json-output is enabled.

  --json-write-ld-context: Write the CURIE map in a `@context` key,
  rather than in a `curie_map` key. This is for compatibility with
  SSSOM-Py. That option is only meaningful when --json-short-iris is
  enabled (otherwise, there is no CURIE map to write).

For convenience, another option, `--sssompy-json`, can be used to
trigger the output of a JSON file that is compatible with SSSOM-Py. The
use of that single option is equivalent to the use of all three options
above combined.