cli: Use positional arguments for input files.
We change the way the CLI tool is called. Input files are now specified
as positional arguments. Any non-option argument on the command line is
assumed to be an input file.

This is for consistency with traditional Unix filter commands (e.g.
cat).

The use of --input options is still completely supported for backwards
compatibility (so, no existing call to sssom-cli will be broken by this
change), but no longer advertised (the option is no longer listed in the
online help message).