Add support for sssom:NoTermFound.
The special value `sssom:NoTermFound` can be used to express that a
given entity in a source domain cannot be mapped to any entity in a
target domain.

The use of that special value is mostly transparent to the SSSOM
implementation (it's up to users to do something with it), apart on one
aspect: mappings that make use of that value (whether on the subject
side or on the object side) must be ignored when we infer cardinality
values.

Consider the following two mappings:

  ORGENT:0001    skos:closeMatch    COMENT:0002
  ORGENT:0001    skos:exactMatch    sssom:NoTermFound

which conjointly state that ORGENT:0001 has a close match in the object
domain, but does not have an exact match in that same domain.

It would not be correct to state that ORGENT:0001 is mapped to more than
one entity (and therefore that both those mappings have a 1:n
cardinality), since the second mapping expresses, on the contrary, an
_absence_ of mapping. So we must ignore the second mapping when
inferring cardinality; the first mapping is a 1:1 mapping, and the
second simply does not have a cardinality at all (a cardinality for a
"non-mapping" would not make any sense).