Move compatibility preprocessors to a distinct package.
We create a new package `org.incenp.obofoundry.sssom.compatibility` and
put all the YAML compatibility preprocessors into it. This is to avoid
cluttering the main package.

This is, strictly speaking, a breaking change, but those preprocessors
have never been intended to be used outside of SSSOM-Core anyway, so I
am not concerned about that.