Remove custom addition of extension_definitions.
The extension_definitions slot will soon be officially added to the
SSSOM schema, so we will no longer need to add that slot to the
MappingSet class as a tweak in our template.

We will _not_ use the auto-generated ExtensionDefinition Java class,
though, as we already have a manually crafted one that contains
additional code.