Update for curie_map being added to the model.
The `curie_map` is or will soon be officially added to the SSSOM schema,
so we will no longer need to manually add that slot to the MappingSet
class.

However, the joke that is the LinkML Java code generator is not able to
recognise that it should be generated as a Map<String,String> (despite
the proper `multivalued`, `inlined`, and `key` attributes), so it
generates it as a List<Prefix> instead. We have to override that.

It seems that support for inlining is a feature of the Python-only
LinkML runtime, _not_ of LinkML in itself. But tell me again that LinkML
is "definitely not only for Python developers"...