Adjust cardinality computation to account for literal mappings.
When computing the cardinality of a list of mappings, we must take into
account that literal mappings do not have their subject_id or object_id
slot (or both); for those mappings, we must the subject_label and/or the
object_label slot.

While we are it, we also take into account the entity_type when
computing the cardinality: for example, if two mappings have the same
subject ID but with a different entity type (probably unlikely to
happen, but who knows), their subjects should be considered different.

We do that by comparing, not only the subject and object IDs/labels, but
a value constructed by prefixing to the ID or label a "tag" representing
the entity type. The tag uses null bytes (\0), so there should be zero
risks of it naturally occuring at the beginning of a legitimate ID or
label.