Validate mappings after parsing.
Update the parsers to include a post-parsing validation step, during
which we check that all the slots that are required by the specification
are present. If that validation step is enabled (it is by default), the
parsers will throw a SSSOMFormatException if the set contains any
mapping that is missing a required slot.