Support the old "literal profile".
For a while, literal mappings were intended to be represented in SSSOM
using a distinct "profile". That profile has never been fully specified
and therefore has never been usable, but nonetheless, some users had
already started crafting "SSSOM/TSV" files with that profile.

We should support those, so we add a new compatibility preprocessor to
convert the "literal*" fields from the literal profile to the new way of
representing literal mappings in SSSOM 1.0.