Allow filtering of literal mappings.
Now that literal mappings are supported, we should allow filtering on
them.

This necessitates two things:

First, the SSSOM/T processor needs updating because it is assuming that
all mappings have a subject_id and an object_id (when it checks whether
the subject or object of a mapping has changed following a preprocessing
operation, so that it knows it should re-compute the cardinality
values).

We fix that by making the processor reuse the same logic as the one
already used by the cardinality computer.

Second, we must make it possible to create subject_label== or
object_label== filters in SSSOM/T.

For completeness, we also allow filtering on predicate_label, since this
is the last text slot that did not have an associated filter. Though of
course filtering on the label rather than the identifier remains a Very
Bad Idea™.