Accept date and time objects when a date is expected.
The SSSOM specification defines `mapping_date` as having a `linkml:Date`
range, which is itself clearly defined as representing only a date
(year, month, day) -- that type is basically the same thing as the XML
Schema type `xsd:date`.

However there are mapping sets in the wild that contain date and time
values in that slot (e.g., "2024-08-05T00:02:15"). For convenience, we
should accept such values, though we truncate it to keep only the date
part.