Remove assumptions that all mappings have a subject/object ID.
There are a few places in the code where we assume that we can simply
call Mapping.getSubjectId() or getObjectId() without worrying about
getting a null reply. That assumption is no longer warranted with
literal mappings.

We fix those assumptions in:

* SSSOM-CLI's "split" feature: we simply exclude all literal mappings
  when splitting a set along the subject and object prefix names;

* the `--only-subject-in` option of the `inject` ROBOT command: by
  definition this option is only meaningful if the subject is not a
  literal, so its use automatically excludes mappings with a literal
  subject.

* likewise for the `--check-subject` option, in the same command;

* likewise, mutatis mutandis, for the `--only-object-in` and
  `--check-object` options.

There may be a few other places where the assumption is made. I am
afraid we will discover them when we'll run into a bug...