Allow filtering on empty slots.
We amend the SSSOM/T filtering logic to allow selecting mappings that do
_not_ have a value for a given slot.

That is, it is now possible to do this:

  mapping_tool=="" -> action();

and the action will apply to all mappings that have no value (either
null, or an empty string) in their mapping_tool slot.

This is currently only possible for filters on text slots, but we will
soon generalise that to other filter types.

closes #5