Allow filtering on empty ID slots.
Make it possible to select mappings based on the absence of value in an
identifier slot. We cannot unfortunately use the obvious =="" syntax (as
we do for text-typed slots), because that would necessitate overhauling
the entire grammar.

Instead, we introduce the special value '~', to be accepted only on
filter that expect identifiers, to signify the absence of value.

That is, a rule like this:

  subject==~ -> action()

would apply the action to all mappings that do not have a subject_id.