Allow filtering on empty lists.
Make it possible to select mappings that have no value in a multi-valued
slot.

This is basically just applying the previous changes to the case where
the slot is a multi-valued slot. For non-identifiers slots, the empty
value is selected by ==""; for identifiers slots, the empty value is
selected by ==~.

For example:

  author==~

would select all mappings that have no values at all in the author_id
slot.