Allow filtering on empty subject_type/object_type.
Since the subject_type and object_type filters accept quoted strings as
values, it is easy to support filtering on empty values for those slots
using the same syntax as for free-form text slots. We do that here.

We also update the documentation to mention the possibility to filter on
empty values for most slots.