Allow filtering on empty numeric slots and cardinality.
Implement the possibility of selecting mappings that do not have a value
in one of the numeric slots (confidence, similarity_score) and in the
mapping_cardinality slot.

With that commit, all slots can be filtered on empty values, except
predicate_modifier. That last slot accepts only one possible value, so
if anyone need to select mappings that do not have a predicate modifier,
it is enough to do

  !predicate_modifier==Not

so there's no need to explicitly allow selecting empty values. We may
revisit that in the future if the predicate_modifier enum gets more
values in future versions of the SSSOM specification.