Remove the intermediate representation of rules.
Parsing a SSSOM/T ruleset was done in 3 steps:

(1) Parse the text into an abstract tree (done by ANTLR4).

(2) Parse the abstract tree into a list of SSSOMTransformRule, an
intermediate representation containing the final filters for each rule,
but not the preprocessors or generators.

(3) Transform each SSSOMTransformRule into a final
MappingProcessingRule, by calling the ISSSOMTransformApplication object
(given to the parser by the client code) to obtain the preprocessors and
generators.

The reason for separating steps (2) and (3) was to "isolate" the code
that had to call the ISSSOMTransformApplication object in a small
portion of the parser code.

However I now believe this is a needless complication which brings
little, if any, benefits.

So here, we remove the intermediate SSSOMTransformRule representation
and basically merge steps (2) and (3) into a single step that parses the
abstract tree and turn it into the final list of MappingProcessingRule,
delegating to the ISSSOMTransformApplication object whenever needed.