Do not offer to expand CURIEs in text arguments.
We remove the getCurieExpansionFormat() method from the
ISSSOMTransformApplication interface, and with it the CURIE expansion
feature that allowed to use CURIEs in the middle of a text argument like
this:

  my_action("something something BFO:0000050 something")

and to get them expanded into their corresponding full-length form when
the function is actually called.

This has never been used in SSSOM-Java itself, whether in SSSOM-CLI or
SSSOM-Robot. The only way to use that feature is to create your own
implementation of ISSSOMTransformApplication. And the feature has never
been documenting elsewhere than in the Javadocs comment of the
ISSSOMTransformApplication interface, so while it it technically a
breaking change, I consider it a minor one.

The feature has always been dubious anyway. It relies on a simple regex
to find CURIEs in the middle of text arguments, which is susceptible to
false positive (finding a CURIE where there is none).