Add application-specific filters.
Allow a SSSOM/T rule to contain a function call within the filter
section, as in:

  subject==ORGENT:* && app_filter("arg1", "arg2") -> action();

where `app_filter` is defined by the SSSOM/Transform application to
perform some custom, application-specific filtering (e.g. filtering that
depends on information not found in the mapping set, but that the
application is aware of).

The ISSSOMTransformApplication interface gets a new member method,
`onFilter`, which is called when a function call is found when parsing
the filters.

Calling application-specific code when parsing the filters introduces
the possibility that the filter-parsing code may now _fail_ to return a
filter (e.g. if the application does not recognise the filter name),
something which was not previously possible. So we need a few guardrails
against null pointers in the ParseTree2FilterVisitor.