Check for unknown functions in the parser.
Do not rely on the ISSSOMTransformApplication implementation to throw a
SSSOMTransformError upon encountering an unknown function. Assume
instead that the application will simply return null (or false, in the
case of "header action" function) if the function name is not
recognised, and will only throw an exception if an actual error occurs
when processing a (recognised) function.

This means the parser should take care of checking whether the calls to
ISSSOMTransformApplication return null/false, and be prepared to deal
with such return values when they occur.