Rename header actions to "directives".
"Header actions" has never been a really good name to refer to those
actions that are not associated with a filter. We rename them to
"directives" instead.

Their main purpose is to influence the behaviour of the application
before parsing is even finished (e.g. by declaring stuff that the
application needs to know in order to be able to process the other kind
of actions), so "directive" seems quite suitable.