Update documentation.
Update the documentation to (1) briefly mention the existence of
application-specific filters and (2) reflect the renaming of "header
actions" to "directives".

The SSSOM/Transform page will likely need to be overhauled to clearly
expose the concept of a SSSOM/T "dialect" and the various types of
functions, but that will have to come later.