Allow defining custom callback functions for SSSOM/T processing.
Allow a SSSOM/Transform application to define completely custom
processing functions, hereafter calling processing callback function or
simply callback functions.

A mapping processing rule can now have an optional callback function. If
it has, said function is invoked by the mapping processor when it is
processing the rule, after checking that the rule should be processed
(i.e. the rule has not been excluded by a tag) but before iterating over
the mappings. When it is invoked, the callback is given two arguments:
the filter associated with the rule (if any), and the current set of
mappings. The callback function can then perform any arbitrary
processing on the mappings (likely using the filter to select the
mappings to process, but not necessarily) and/or change the state of the
application to induce side-effects.

A SSSOM/Transform application can define a callback function by
implementing the onCallback(String, List<String>) method, which is
called whenever a function call is encountered when parsing the SSSOM/T
rules.