Use new features to implement SSSOM/T-OWL dialect.
Now that we can have application-specific filters and custom callback
methods, we can use them in SSSOM/T-OWL.

Specifically, we add two filters that can be used to replace the
`check_subject_existence()` and `check_object_existence()`
preprocessors: `subject_exists()` and `object_exists()`. For example:

  subject==FBbt:* -> check_subject_existence();

can be replaced by:

  subject==FBbt:* && !subject_exists() -> stop();

We also add two filters to select mappings depending on whether the
subject or object is a descendant of a given class in the ontology:
`subject_is_a(CLASS)` and `object_is_a(CLASS)`.

Finally, we use callback functions to allow defining mapping-dependent
variables using the filter syntax, as in:

  subject==FBbt:* -> set_var("MY_VAR", "FBbt-specific value");

This means we can replace a construct like this:

  set_var("TAXREL", BFO:0000066, "%object_id is_a UBERON:0000105");

(which has always been WEIRD) by something much more elegant (IMHO):

  object_is_a(UBERON:0000105) -> set_var("TAXREL", BFO:0000066);

The old constructs (`check_subject_existence`, `check_object_existence`,
and the 3-arguments form of `set_var`) are still supported but are
deprecated and no longer mentioned in the documentation. They will be
removed in a future version.