Avoid the use of MappingFormatter.format().
The `format(String, Mapping)` method of the MappingFormatter object is
going to be deprecated (or, at the very least, its use is going to be
discouraged), in favour of using the transformer object returned by
`getTransformer(String)`.

To prepare for this, we udpdate the only two places where that method
was used, to use a transformer object instead.