Overhaul the MappingFormatter.
Completely re-write the MappingFormatter class.

The formatter is now designed to recognise "bracketed" placeholders in a
format string, of the form "%{placeholder|modifier}" (the "|modifier"
bit is optional).

The logic of the new implementation is reversed compared to the
precedent iteration. In the initial code, the placeholders were searched
for and replaced in the `format(String, Mapping)` method, when it was
time to actually format a string; and the `getTransformer(String)`
method merely returned a lambda that internally called the `format()`
method when the transformer was applied. This meant that the format
string had to be searched for placeholders every single time the
transformer was applied.

In the new implementation, the format string is parsed once in the
`getTransformer(String)` method, which returns a transformer that
contains a sort of "compiled" version of the format string, and does not
require any more parsing when the transformer is applied; and the
`format(String, Mapping)` method works by internally calling
`getTransformer()` and immediately applying the transformer. In
addition, `format()` keeps a cache of the transformer objects it obtains
in the lifetime of the MappingTransformer, so that repeated calls to
`format()` with the same format string will not require parsing the
format string again.

The parser is relatively laxist and will silently ignore most cases that
could be considered errors, such as a '%' sign followed by something
else than a letter character. It will only error out (with a
IllegalArgumentException) when the format string ends on a
non-terminated bracketed placeholder (like "%{place").

Though this will not be advertised in the documentation, the formatter
still supports the previous mode (used by SSSOM/T-OWL) where
placeholders were recognised when preceded by a single '%' character (as
in "%placeholder"). Such "legacy placeholders" are intended for
backwards compatibility with earlier versions of SSSOM/T-OWL only.