Allow the use of "naked" placeholders in arguments.
If the argument of a function is a single placeholder, allow to specify
that argument without enclosing it in quotes.

That is, a call like this:

  function("%{subject_id}");

can be written like this:

  function(%{subject_id});

This includes the possibility of using a modifier:

  function(%{subject_id|modifier})

However when not enclosed within quotes, the placeholder and the
modifier can only contains ASCII letters, numbers, and underscores.