Add a full test using the Uberon-FBbt bridge.
To test the sssom:inject command, we use a SSSOM/T-OWL ruleset that
is much closer to the actual ruleset used in Uberon, and we also use the
full FBbt and FBdv mapping sets as input, as well as a full copy of the
latest Uberon release.

This makes the test slightly longer, but also makes it much more useful
as if that test passes, I can be fairly confident the Uberon bridging
pipeline would not be broken by any changes brought to the code.