Add keyed arguments to SSSOM/T.
Add the possibility of passing "keyed arguments" to a SSSOM/T function.

A keyed argument has the form "/key=VALUE", where VALUE can be any other
kind of argument accepted by a function (a quoted string, a IRI, a
CURIE, or a placeholder).

The ISSSOMTransformApplication interface is updated so that all the
methods that are called upon encountering a function call receive an
extra "keyedArguments" parameter, containing a map with all the keyed
arguments found in the call.