Remove the "fast path" for create_axiom.
The idea behind the "fast path" in the implementation of SSSOM/T-OWL's
create_axiom was that for commonly used axioms, we could recognise them
for what they were and directly craft the axiom by using the
OWLDataFactory rather than having to rely on the Manchester parser. In
theory, this should be faster, since we do not have to parse the
Manchester expression for every single mapping for which create_axiom is
called.

In practice, when working with Uberon's mapping set and bridging rules,
it was found that the "fast path" was actually _not_ significantly
faster than the normal path, so the only thing the "fast path" does is
making the code of the SSSOM/T-OWL application more complex for no
discernable benefits.

"Kill your darlings", they say. That's what we do here. The "fast path"
was an ill-conceived idea that did not respond to any need. A case of
"premature optimisation".