Model SSSOM/T functions as classes.
This commit introduces a new implementation of the
ISSSOMTransformApplication interface in which SSSOM/T functions are
represented each by an individual Java class. It is intended to replace
the SSSOMTransformApplicationBase implementation.

The new approach is somewhat more verbose, but I believe it is
ultimately clearer, as the code for each function is in a dedicated
class rather than in the middle of a switch/case block.

The new implementation sits aside the old one, for now. When all
functions (including the SSSOM/T-OWL functions) will have been
re-implemented using the new model, SSSOMTransformApplicationBase will
be removed.