Allow substitutions in SSSOM/T assign arguments.
Make the SSSOM/T "assign" function accept placeholders in its second
argument, so that the value that is assigned to a slot can be partially
or totally derived from the very mapping that is being modified (using
some values from other slots, or a mapping-dependent variable).

We do _not_ extend that feature to the "edit" function, which is
deprecated. We also do not extend it to the "replace" function, where it
would probably not be worth the additional complexity.