Migrate SSSOM/T-OWL to the new SSSOM/T application model.
This is a huge commit, but it is actually pretty straightforward.

We add a new class SSSOMTOwlApplication, derived from the base
SSSOMTransformApplication. It will contain the state of the application,
and provide some helper methods for the benefits of the SSSOM/T-OWL
function. It also enables the "unbracketed" substitutions that were
specific to the previous SSSOM/T-OWL implementation. For clarity, the
CustomEntityChecker, which in the previous implementation is an inner
class, is here moved to a class of its own.

Then we add a new class for each of the SSSOM/T-OWL functions. While we
are at it, we add a few functions intended to replace some of the old
functions:

* "declare", as a replacement for both "declare_class" and
  "declare_object_property";

* "exists", as a replacement for both "subject_exists" and
  "object_exists";

* "is_a", as a replacement for both "subject_is_a" and "object_is_a";

* "annotate", as a replacement for both "annotate_subject" and
  "annotate_object".

The old functions are still supported for backwards compatibility,
except the "*_exists" and "*_is_a" functions: those were never part of
any released version of SSSOM/T-OWL (they had been recently added to the
main branch, *after* the last release), so there is no need to keep them
around.

We update the "fbbt-bridge.rules" test file to make it use the new
functions.

Lastly, in SSSOM-Robot the only thing to do is to simply replace the old
SSSOMTOwl object by an instance of the new SSSOMTOwlApplication.