Allow to pass arguments to substitution modifiers.
Make it possible to pass arguments to a substitution modifier, as in:

  %{placeholder|modifier(arg1, arg2)}

The modifier function will then be called with 3 arguments: the value of
the placeholder, and the strings "arg1" and "arg2".

A modifier function is now represented as a ISSSOMTFunction<String>
rather than as a Function<String,String>, to account for the possibility
of passing more than one argument.

closes #9