Support chaining modifier functions in a format string.
Allow the use of several modifier functions applied to the same
placeholder, as in:

  %{placeholder|modifier1|modifier2|modifier3}

Each function is given as input the output of the previous function (or
the placeholder value itself, for the first function). In the example
above, `modifier3` is called with the output of `modifier2`, which it
iself called with the output of `modifier1`, which is called with the
value of `placeholder`.

Each function can have its own additional arguments, e.g.:

  %{placeholder|modifier1|modifier2_with_args(my_arg)|modifier3}

While adding this feature, we refactor the format string parser. We now
use a handful of helper internal classes to represent a "compiled" form
of the format string and its individual components. This makes the
parsing code much more readable, in part because it also reduces the
number of lambda functions.