Make the 'assign' function more predictable.
When the SSSOM/T 'assign' function is called to edit more than one field
in a single step, the actual result of the entire operation can be
difficult to predict because of the order in which the changes are
applied (that is, the order in which the slots are visited). That order
is deterministic, but unknown to users.

For example, consider the following call:

  assign('subject_id', <some_new_id_value>,
         'comment', 'new name of %{subject_id}');

The value assigned to the `comment` slot will depend on whether the slot
is visited before or after the `subject_id` slot. It so happens that
`subject_id` is visited first, so by the time we set the value of the
`comment` slot, `subject_id` already contains the new value.

To avoid this kind of situation, we keep a copy of the original mapping
for the duration of the entire 'assign' operation, and always use that
copy as the source for all interpolations.