Expand URI Expression support to SSSOM/T-OWL.
Add a new substitution modifier function to SSSOM/T-OWL to automatically
expand a URI Expression into a pre-defined OWL class expression based on
a known per-schema template.

For now, this requires that the per-schema templates are explicitly
registered as part of the SSSOM/T-OWL ruleset using the new directive
function `uriexpr_declare_format`. Other out-of-band mechanisms to
register a schema may be added later.