Allow modifier functions to work on any type of objects.
Change the way modifier functions work by not making them receive and
return strings, but objects.

This is to make it possible for a modifier function to work on the
original mapping-derived value, without a pre-established conversion to
string.

For example, a function called after a placeholder that emits the value
of the mapping_date slot could work on the LocalDate object itself,
rather than on a string representation of the date object.

This requires changing the way modifier functions are represented, using
a new interface specifically for them.