Accept quoted arguments in modifier calls.
It is annoying that arguments in a modifier call cannot contain a comma
(which is always interpreted as the argument separator) or that
whitespaces are always trimmed (forbidding having an argument consisting
only of whitespaces).

So, let's make the parser for formatting strings *slightly* more complex
by recognising three different types of modifier arguments: unquoted (as
before), quoted with single quotes, and quoted with double quotes.

We recognise the two types of quotes so that the user can use inside a
modifier call the type of quote that she has not been using to quote the
entire format string, as in:

  "here is a placeholder: %{placeholder|modifier('arg1', 'arg2')}"