Allow the use of a default modifier with legacy substitutions.
When parsing a format string, if a legacy placeholder is found, allow to
apply to it a default modifier function.

We use that feature in SSSOM/T-OWL's `create_axiom` function so that any
legacy placeholder found in its argument will be rendered between angled
brackets. That is, this:

  create_axiom("%subject_id SubClassOf: %object_id");

is equivalent to

  create_axiom("%{subject_id|format(<%s>)} SubClassOf: %{object_id|format(<%s>)}");

which of course is also equivalent to:

  create_axiom("<%{subject_id}> SubClassOf: <%{object_id}>");

The idea is that "legacy placeholders" now becomes "placeholders that
may be treated differently between different functions so as to be the
most useful".