Remove the concept of "legacy placeholders".
Now that we have function-specific default modifiers for un-bracketed
placeholders (allowing us to preserve the old behaviour of SSSOM/T-OWL's
create_axiom() function), we no longer need to maintain different list
of placeholders depending on whether they should be recognised with or
without brackets.

All defined substitutions are now automatically usable with both the
bracketed and the un-bracketed syntax (provided the placeholder name is
compatible with the un-bracketed syntax, that is, it does not contain
anything else than letters and underscores).